+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Releases
 
-## Version 4.1.9 20250801 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+## Version 4.1.10 So Far... [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+
+- "Must follow block" martial arts response element no longer behaves as a block maneuver. [#2528](https://github.com/dmdorman/hero6e-foundryvtt/issues/2528)
+
+## Version 4.1.9 20250801
 
 - Martial arts with weapons can now do killing damage. [#2531](https://github.com/dmdorman/hero6e-foundryvtt/issues/2531)